Rename test library to be qtest
