+++ pg/pg_test.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/shiblon/entroq"
 	grpcbackend "github.com/shiblon/entroq/grpc"
-	"github.com/shiblon/entroq/qsvc/test"
+	"github.com/shiblon/entroq/qsvc/qtest"
 	"golang.org/x/sync/errgroup"
 	"google.golang.org/grpc"
 
@@ -109,7 +109,7 @@ func TestSimpleSequence(t *testing.T) {
 
 	pgHostPort := fmt.Sprintf("localhost:%v", pgPort)
 
-	server, dial, err := test.StartService(ctx, Opener(pgHostPort,
+	server, dial, err := qtest.StartService(ctx, Opener(pgHostPort,
 		WithDB("postgres"),
 		WithUsername("postgres"),
 		WithPassword("password")))
@@ -128,5 +128,5 @@ func TestSimpleSequence(t *testing.T) {
 	}
 	defer client.Close()
 
-	test.SimpleSequence(ctx, t, client)
+	qtest.SimpleSequence(ctx, t, client)
 }
